requirements.txt
absl-py==0.9.0
adaptnlp==0.1.3
alabaster==0.7.12
altair==4.1.0
anaconda-client==1.7.2
anaconda-navigator==1.8.7
anaconda-project==0.8.3
analytics-python==1.2.9
appdirs==1.4.3
argh==0.26.2
asn1crypto==1.3.0
astor==0.8.0
astroid==2.3.3
astropy==4.0
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
awscli==1.17.8
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bamboolib==1.7.0
base58==2.0.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bitarray==1.2.1
bkcharts==0.2
black==19.10b0
bleach==3.1.0
blinker==1.4
blis==0.4.1
bokeh==1.4.0
boto==2.49.0
boto3==1.12.12
botocore==1.15.12
Bottleneck==1.3.2
bpemb==0.3.0
brunel==2.6.2
bz2file==0.98
cachetools==4.0.0
camelot-py==0.7.3
catalogue==1.0.0
cchardet==2.1.5
cdqa==1.3.9
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.4.3
comtypes==1.1.7
conda==4.8.3
conda-build==3.10.5
conda-package-handling==1.6.0
conda-verify==3.4.2
conllu==2.2.1
contextlib2==0.6.0.post1
cplex==12.10.0.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.2
Cython==0.29.15
cytoolz==0.10.1
dash==1.7.0
dash-bootstrap-components==0.7.2
dash-core-components==1.6.0
dash-html-components==1.0.2
dash-renderer==1.2.2
dash-table==4.5.1
dask==2.11.0
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.7
diff-match-patch==20181111
dill==0.2.9
distlib==0.3.0
distributed==2.12.0
distro==1.5.0
docloud==1.0.375
docplex==2.12.182
docutils==0.16
en-core-sci-lg==0.2.4
en-core-sci-sm==0.2.4
en-core-web-md==2.2.5
en-core-web-sm==2.2.0
en-ner-bc5cdr-md==0.2.4
en-ner-bionlp13cg-md==0.2.4
entrypoints==0.3
enum-compat==0.0.3
environment-kernels==1.1.1
et-xmlfile==1.0.1
fastcache==1.1.0
fastprogress==0.2.2
filelock==3.0.12
flair==0.4.5
flake8==3.7.9
flashtext==2.7
Flask==1.1.1
Flask-Compress==1.4.0
Flask-Cors==3.0.8
fsspec==0.6.2
future==0.18.2
fuzzywuzzy==0.17.0
gast==0.2.2
gdown==3.10.1
gensim==3.8.1
gevent==1.4.0
ghostscript==0.6
glob2==0.7
gluonnlp==0.9.1
google==2.0.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
google-search==1.0.2
googleapis-common-protos==1.51.0
googledrivedownloader==0.4
graphviz==0.8.4
greenlet==0.4.15
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
HeapDict==1.0.1
html-text==0.5.1
html2text==2019.9.26
html5lib==1.0.1
httplib2==0.17.0
hyperopt==0.2.3
hypothesis==5.5.4
idna==2.9
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==1.5.0
importlib-resources==1.3.1
intervaltree==3.0.2
ipykernel==5.1.4
ipympl==0.5.3
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
jieba==0.42.1
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
json5==0.9.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-pip==0.3.1
jupyterlab==1.2.6
jupyterlab-server==1.0.7
Keras==2.3.1
Keras-Applications==1.0.8
keras-bert==0.81.0
keras-embed-sim==0.7.0
keras-layer-normalization==0.14.0
keras-multi-head==0.22.0
keras-pos-embd==0.11.0
keras-position-wise-feed-forward==0.6.0
Keras-Preprocessing==1.1.0
keras-self-attention==0.41.0
keras-transformer==0.32.0
keyring==21.1.0
kiwisolver==1.1.0
ktrain==0.10.1
langdetect==1.0.7
lazy-object-proxy==1.4.3
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
matplotlib-venn==0.11.5
mccabe==0.6.1
medacy==0.2.0
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.1
more-itertools==8.2.0
mpld3==0.3
mpmath==1.1.0
msgpack==0.6.1
msgpack-numpy==0.4.3.2
multipledispatch==0.6.0
multiprocess==0.70.9
murmurhash==1.0.2
mxnet==1.6.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
nmslib==2.0.5
nose==1.3.7
notebook==6.0.3
npm==0.1.1
numba==0.48.0
numexpr==2.7.1
numpy==1.18.2
numpydoc==0.9.2
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.2.30
openpyxl==3.0.3
opt-einsum==3.2.0
optional-django==0.1.0
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathos==0.2.5
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.1
pdfminer.six==20200104
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
plac==0.9.6
plotly==4.5.2
pluggy==0.13.1
ply==3.11
pox==0.2.7
ppft==1.6.6.1
preshed==3.0.2
prettytable==0.7.2
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.3
protobuf==3.11.4
psutil==5.7.0
psycopg2==2.8.5
PuLP==2.0
py==1.8.1
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.4.3
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.4
pycurl==7.43.0.5
pydeck==0.3.0b4
pydocstyle==4.0.1
PyDrive==1.3.1
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
PyMuPDF==1.16.10
PyNaCl==1.3.0
pyodbc===4.0.0-unsupported
Pyomo==5.6.8
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyPDF2==1.26.0
pyreadline==2.1
pyrsistent==0.15.7
pysbd==0.2.3
PySocks==1.7.1
pytest==5.4.1
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-crfsuite==0.9.6
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
pytorch-crf==0.7.2
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.2.0
pytz==2019.3
PyUtilib==5.7.3
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyxlsb==1.0.6
PyYAML==5.3
pyzmq==18.1.1
QDarkStyle==2.8
qgrid==1.3.0
QtAwesome==0.7.0
qtconsole==4.7.1
QtPy==1.9.0
rank-bm25==0.2
readme-renderer==25.0
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rope==0.16.0
rsa==3.4.2
Rtree==0.9.3
ruamel-yaml==0.15.87
s3transfer==0.3.3
sacremoses==0.0.38
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
scispacy==0.2.4
seaborn==0.10.0
segtok==1.5.7
Send2Trash==1.5.0
sentencepiece==0.1.85
seqeval==0.0.12
simplegeneric==0.8.1
simpletransformers==0.20.3
singledispatch==3.4.0.3
six==1.14.0
sklearn-crfsuite==0.3.6
smart-open==1.9.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==2.0
spacy==2.2.3
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.0
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.15
sqlitedict==1.6.0
srsly==1.0.1
statsmodels==0.11.0
streamlit==0.57.3
sympy==1.5.1
syntok==1.2.2
tables==3.6.1
tabula-py==2.1.0
tabulate==0.8.6
tblib==1.6.0
tensorboard==2.1.0
tensorboardX==2.0
tensorflow==1.15.0
tensorflow-datasets==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu-estimator==2.1.0
tensorflow-metadata==0.21.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textdistance==4.1.5
thinc==7.3.1
tika==1.19
tokenizers==0.5.2
toml==0.10.0
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.44.1
traitlets==4.3.3
transformers==2.5.1
twine==3.1.1
typed-ast==1.4.1
tzlocal==2.0.0
ujson==1.35
unicode==2.7
unicodecsv==0.14.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
validators==0.14.3
virtualenv==20.0.10
wasabi==0.6.0
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
wget==3.2
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.6.0
wrapt==1.12.1
xgboost==0.90
xlrd==1.2.0
XlsxWriter==1.2.8
xlwings==0.18.0
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.28.0
zict==2.0.0
zipp==3.1.0
